+++ cjs/lib/__bundle-a9ae393e.cjs
@@ -1,0 +1,1 @@
+var e=require("./__bundle-a6c826ee.cjs");class s extends e.User{constructor(e,s){var t;super(e,s),this.isMuted=!1,this.isMuted=null!==(t=s.is_muted)&&void 0!==t&&t}}exports.Participant=s;